Add lasagna, pizza recipes + back button + references
This commit adds two new recipes to the website: lasagna and pizza.
It also adds a back button to the recipe pages, so that users can easily navigate back to the main page.
Updated references in index.html to include the new recipes orderd in a list.
Refer to theodinproject.com / Project: Recipe / Iteration 4 for more information.